Refactor uncruddy GymMnagaersController method to be cruddy
